Move spanKind to only use
The spanKind func is used by `tracer.Start` and nowhere else. Move its
definition closer to its use and similarly adjust the tests.
