+++ sdk/trace.go
@@ -21,22 +21,6 @@ import (
 	"go.opentelemetry.io/auto/sdk/internal/telemetry"
 )
 
-func spanKind(kind trace.SpanKind) telemetry.SpanKind {
-	switch kind {
-	case trace.SpanKindInternal:
-		return telemetry.SpanKindInternal
-	case trace.SpanKindServer:
-		return telemetry.SpanKindServer
-	case trace.SpanKindClient:
-		return telemetry.SpanKindClient
-	case trace.SpanKindProducer:
-		return telemetry.SpanKindProducer
-	case trace.SpanKindConsumer:
-		return telemetry.SpanKindConsumer
-	}
-	return telemetry.SpanKind(0) // undefined.
-}
-
 type span struct {
 	noop.Span
